+++ src/main/java/movwe/services/authServices/LoginRequestService.java
@@ -1,4 +1,4 @@
-package movwe.services.moderatorServices;
+package movwe.services.authServices;
 
 import lombok.AllArgsConstructor;
 import movwe.domains.logins.LoginRequest;
